requirements.txt
-f https://download.pytorch.org/whl/cu113
aiosqlite==0.17.0
albumentations==1.1.0
databases==0.5.5
fastapi-socketio==0.0.9
opencv-python==4.5.5.64
pandas==1.3.5
pydantic==1.9.1
pysqlite3==0.4.7
pytorch_lightning==1.6.3
scikit-learn==1.0.2
seaborn==0.11.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
uvicorn==0.17.6
wandb==0.12.16
